ml_engine/requirements.txt
torch==2.1.0
torchvision==0.16.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
timm==0.9.12
albumentations==1.3.1
scikit-learn==1.3.2
numpy==1.24.4
python-dotenv==1.0.0
